create writer interface to existing GEDFile class
